% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird.R
\docType{data}
\name{bird}
\alias{bird}
\title{Phylogenetic data files recording body mass, brain volume and brain mass: birds}
\format{
A data frame with 10176 rows and 20 variables
\describe{
\item{Order}{Order}
\item{Family}{Family}
\item{Genus}{Genus}
\item{Species}{Species}
\item{Subspecies}{Subspecies}
\item{Genus_Species}{Genus species}
\item{Sex}{Gender: m-male, f-female}
\item{Sample size}{The sample size used to estimate static allometries}
\item{Brain Volume (ml)}{Brain volume}
\item{Method}{Method of measuring brain size}
\item{Brain mass (g)}{Brain mass in grams}
\item{Body mass (g)}{Body mass in grams}
\item{Specimen ID}{Specimen ID}
\item{Museum}{Museum}
\item{Date (mm-dd-yyyy)}{Date}
\item{Locality}{Locality}
\item{Coordinate}{Coordinate}
\item{Age Class}{Age class}
\item{Reference}{Reference}
\item{Remark}{Remark}
}
}
\source{
Tsuboi, M., van der Bijl, W., Kopperud, B.T., Erritzøe, J., Voje, K.L., Kotrschal, A., Yopak, K.E., Collin, S.P., Iwaniuk, A.N. and Kolm, N., 2018. Breakdown of brain–body allometry and the encephalization of birds and mammals. Nature Ecology & Evolution, 2(9), pp.1492-1500.
}
\usage{
bird
}
\description{
Phylogenetic data files recording body mass, brain volume and brain mass of birds by Tsuboi, M., van der Bijl, W., Kopperud, B.T., Erritzøe, J., Voje, K.L., Kotrschal, A., Yopak, K.E., Collin, S.P., Iwaniuk, A.N. and Kolm, N., 2018. Breakdown of brain–body allometry and the encephalization of birds and mammals. Nature Ecology & Evolution, 2(9), pp.1492-1500.
}
\examples{
data(bird)
}
\keyword{datasets}
